requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
Brotli==1.1.0
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.4
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.0.14
django-debug-toolbar==5.2.0
django-extensions==4.1
djangorestframework==3.16.0
dnspython==2.7.0
eventlet==0.40.1
fonttools==4.58.1
greenlet==3.2.3
gunicorn==23.0.0
html5lib==1.1
idna==3.10
kombu==5.5.3
MarkupSafe==3.0.2
oauthlib==3.3.0
packaging==25.0
pillow==10.3.0
prompt_toolkit==3.0.51
psycopg==3.2.9
psycopg2==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
redis==6.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
stripe==9.3.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
zopfli==0.2.3.post1
